Update _index.md
Since configuring usage billing for Enterprise license requires a contractual change and billing rates, if customers send a support ticket for this, we'd just redirect them to their account team anyways. Support also does not handle changing this as the account team would route this to the Provisioning team for actioning.